Restructure folder and files
